feat: ajustado logica de confirmação de ações de deletar para os componentes de botões padrão da listagem
