+++ components/molecules/Datatable/ZDatatableGeneric.vue
@@ -2,26 +2,33 @@
   <div class="row justify-start mt-3">
     <div class="flex flex-col xs2">
       <div class="item">
-        <ZDataTableActionButtons @add="actionAdd" @delete="actionDeletes" :selectedItemsEmitted="selectedItemsEmitted" />
+        <ZDataTableActionButtons
+          @add="actionAdd"
+          @delete="actionDeletes"
+          :selectedItemsEmitted="selectedItemsEmitted"
+        />
       </div>
     </div>
     <div class="flex flex-col offset-xl7 xs2">
       <div class="item">
-        <ZDataTableInputSearch v-model="search" @input="updateSearch"  />
+        <ZDataTableInputSearch
+          v-model="search"
+          @input="updateSearch"
+        />
       </div>
     </div>
   </div>
-  <ZDataTable 
-    :items="items" 
-    :columns="columns" 
+  <ZDataTable
+    :items="items"
+    :columns="columns"
     :selectable="selectable"
     :includeActionsColumn="includeActionsColumn"
     @selectionChange="selectedItemsEmitted = $event.currentSelectedItems"
   >
     <template #cell(actions)="{ id }">
-      <ZDataTableActions 
+      <ZDataTableActions
         :id="id"
-        @edit="actionEdit" 
+        @edit="actionEdit"
         @delete="actionDelete"
       />
     </template>
@@ -37,7 +44,6 @@ import ZDataTableInputSearch from '~/components/molecules/Datatable/ZDataTableIn
 import ZDataTable from '~/components/molecules/Datatable/ZDataTable'
 import ZInput from '~/components/atoms/Inputs/ZInput';
 import ZIcon from '~/components/atoms/Icons/ZIcon';
-import Swal from 'sweetalert2'
 
 export default defineComponent({
   components: {
@@ -94,45 +100,8 @@ export default defineComponent({
       this.$emit('add');
     },
 
-    actionDeletes() {
-      const itemsDelete = this.selectedItemsEmitted.map((item) =>
-         item.id
-      )
-
-      const totalItems = this.selectedItemsEmitted.length
-
-      Swal.fire({
-        title: 'Deseja deletar estes registros?',
-        text: "Você não será capaz de reverter isso!",
-        icon: 'warning',
-        html:
-          'Você tem <b>' + totalItems + '</b> registros selecionados para deletar.<br>',
-        showCancelButton: true,
-        confirmButtonText: 'Sim, deletar!',
-        confirmButtonColor: '#154EC1',
-        cancelButtonColor: '#E42222',
-        cancelButtonText: 'Não, cancelar!',
-      }).then((result) => {
-        if (result.isConfirmed) {
-          Swal.fire({
-            title: 'Deletando!',
-            text: 'Seus registros estão sendo deletados!',
-            icon: 'info',
-            showConfirmButton: false,
-          })
-          this.$emit('deletes', itemsDelete);
-        } else if (
-          /* Read more about handling dismissals below */
-          result.dismiss === Swal.DismissReason.cancel
-        ) {
-          Swal.fire({
-            title: 'Cancelado',
-            text: 'Seus registros ainda continuam existindo :)',
-            icon: 'error',
-            confirmButtonColor: '#154EC1',
-          })
-        }
-      })
+    actionDeletes(itemsDelete) {
+      this.$emit('deletes', itemsDelete);
     },
 
     actionDelete(id) {
